fix player

+++ database/migrations/2023_12_04_101730_add_places_to_ranking_table.php
@@ -14,7 +14,7 @@ class AddPlacesToRankingTable extends Migration
     public function up()
     {
         Schema::table('ranking', function (Blueprint $table) {
-            $table->integer('places');
+            $table->bigInteger('places');
         });
     }
